medcq/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
firebase-admin==6.1.0
supabase==0.7.1
pytest==7.3.1
httpx==0.24.1
alembic==1.11.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
loguru==0.7.0